+++ src/slt_header.jl
@@ -4,7 +4,7 @@
 # Thünen Institut
 # Institut für Agrarklimaschutz
 # Junior Research Group NITROSPHERE
-# Julia 0.6
+# Julia 0.6.1
 # 09.12.2016
 # Last Edit: 12.05.2017
 
@@ -22,7 +22,7 @@ end # slt_header(f::String,AnalogIn::Int,freq::String)
 
 function slt_header(f::String,AnalogIn::Int,freq::Number)
 	# Prepare Output Variables
-	bpr = Int64 # Bytes Per Record
+	bpr = Int # Bytes Per Record
 	eddymeasver = [] # Eddy Meas Version
 	t0 = DateTime # Initial Timestamp
 	ch = [] # Channels
@@ -33,13 +33,13 @@ function slt_header(f::String,AnalogIn::Int,freq::Number)
 	fid = open(f,"r")
 	try
 		# Header
-		bpr = Int64(read(fid,Int8,1)[1]) # Bytes Per Record
+		bpr = Int(read(fid,Int8,1)[1]) # Bytes Per Record
 		eddymeasver = read(fid,Int8,1)
-		dom = Int64(read(fid,Int8,1)[1]) # Day of Month
-		m = Int64(read(fid,Int8,1)[1]) # Month number
-		yr = 100*Int64(read(fid,Int8,1)[1]) + Int64(read(fid,Int8,1)[1]) # Year
-		h = Int64(read(fid,Int8,1)[1]) # Hour
-		minut = Int64(read(fid,Int8,1)[1]) # Minute
+		dom = Int(read(fid,Int8,1)[1]) # Day of Month
+		m = Int(read(fid,Int8,1)[1]) # Month number
+		yr = 100*Int(read(fid,Int8,1)[1]) + Int(read(fid,Int8,1)[1]) # Year
+		h = Int(read(fid,Int8,1)[1]) # Hour
+		minut = Int(read(fid,Int8,1)[1]) # Minute
 		t0 = DateTime(yr,m,dom,h,minut,0) # Initial file time
 		
 		fs = stat(f).size # File size in bytes
